move components to about folder
