+++ pages/about/for-legislators.tsx
@@ -1,5 +1,5 @@
 import { createPage } from "../../components/page"
-import ForLegislators from "../../components/ForLegislators/ForLegislators"
+import ForLegislators from "../../components/about/ForLegislators/ForLegislators"
 
 export default createPage({
   title: "MAPLE for Legislators",
