+++ pages/about/for-testifiers.tsx
@@ -1,5 +1,5 @@
 import { createPage } from "../../components/page"
-import ForTestifiers from "../../components/ForTestifiers/ForTestifiers"
+import ForTestifiers from "../../components/about/ForTestifiers/ForTestifiers"
 
 export default createPage({
   title: "MAPLE for Testifiers",
